requirements.txt
fastapi
uvicorn
beanie
motor
pydantic
pydantic
python-jose
passlib
python-multipart
gradio
requests
python-dotenv
